requirements.txt
# Deep Learning Framework
tensorflow==2.10.0
keras==2.10.0

# Data Processing
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
pillow==10.0.0

# User Interface
gradio==4.7.1

# Utilities
scikit-learn==1.3.0